#184: toolbars splited into multiple files
