+++ package.json
@@ -17,23 +17,23 @@
     "@reduxjs/toolkit": "^1.5.0",
     "ace-builds": "^1.4.12",
     "dotenv": "8.2.0",
-    "firebase": "8.2.9",
+    "firebase": "8.2.10",
     "firebase-admin": "9.5.0",
     "fp-ts": "^2.9.5",
     "io-ts": "^2.2.15",
-    "konva": "^7.2.4",
-    "next": "^10.0.7",
+    "konva": "^7.2.5",
+    "next": "^10.0.8",
     "nookies": "^2.5.2",
     "react": "^17.0.1",
     "react-ace": "^9.3.0",
     "react-color": "^2.19.3",
     "react-dom": "^17.0.1",
-    "react-firebase-hooks": "^2.2.0",
-    "react-grid-layout": "^1.2.0",
-    "react-hotkeys-hook": "^3.1.0",
+    "react-firebase-hooks": "^3.0.1",
+    "react-grid-layout": "^1.2.2",
+    "react-hotkeys-hook": "^3.2.1",
     "react-konva": "17.0.1-3",
     "react-redux": "^7.2.2",
-    "react-resize-detector": "^6.6.0",
+    "react-resize-detector": "^6.6.1",
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0"
   },
@@ -51,12 +51,12 @@
     "@types/react-redux": "^7.1.16",
     "@types/react-resize-detector": "^5.0.0",
     "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.15.2",
-    "@typescript-eslint/parser": "^4.15.2",
-    "autoprefixer": "^10.2.4",
+    "@typescript-eslint/eslint-plugin": "^4.16.1",
+    "@typescript-eslint/parser": "^4.16.1",
+    "autoprefixer": "^10.2.5",
     "babel-jest": "^26.6.3",
     "enzyme": "^3.11.0",
-    "eslint": "^7.20.0",
+    "eslint": "^7.21.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
@@ -67,9 +67,9 @@
     "husky": "^5.1.1",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
-    "ts-jest": "^26.5.2",
-    "typescript": "^4.2.2",
-    "webpack": "^5.24.2"
+    "ts-jest": "^26.5.3",
+    "typescript": "^4.2.3",
+    "webpack": "^5.24.3"
   },
   "jest": {
     "moduleNameMapper": {
